掛號送出前檢查欄位型態、自費更新(初版)、加藥(初版)

+++ sql/each_use_medicines.sql
@@ -2,7 +2,7 @@ create table each_use_medicines (
 	`no` INT PRIMARY KEY AUTO_INCREMENT,
 	`time` datetime default current_timestamp,
 	`aId` int(11),
-	`quantity` int(11),
+	`quantity` decimal(10,2),
 	`reason` varchar(50),
 	`mark` varchar(150),
     `emId` int(11),
